edited

+++ controllers/admin.js
@@ -1,19 +1,19 @@
+const mongoose = require('mongoose');
+
+const fileHelper = require('../util/file');
+
+const { validationResult } = require('express-validator/check');
+
 const Product = require('../models/product');
-
-
-// const fileHelper = require('../util/file');
-
-const {validationResult}=require('express-validator/check');
-const { response } = require('express');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
     editing: false,
-    hasError:false,
-    errorMessage:null,
-    validationErrors:[],
+    hasError: false,
+    errorMessage: null,
+    validationErrors: []
   });
 };
 
@@ -22,42 +22,44 @@ exports.postAddProduct = (req, res, next) => {
   const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  if(!image){
+  if (!image) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
-      path: '/admin/edit-product',
+      path: '/admin/add-product',
       editing: false,
-      hasError:true,
+      hasError: true,
       product: {
-        title:title,
-        price:price,
-        description:description
+        title: title,
+        price: price,
+        description: description
       },
-      errorMessage:'Attached file is not image.',
-      validationErrors:[],
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: []
     });
   }
-  const errors=validationResult(req);
+  const errors = validationResult(req);
 
-  if(!errors.isEmpty()){
+  if (!errors.isEmpty()) {
+    console.log(errors.array());
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
-      path: '/admin/edit-product',
+      path: '/admin/add-product',
       editing: false,
-      hasError:true,
+      hasError: true,
       product: {
-        title:title,
-        price:price,
-        description:description
+        title: title,
+        price: price,
+        description: description
       },
-      errorMessage:errors.array()[0].msg,
-      validationErrors:errors.array(),
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array()
     });
   }
 
-  const imageUrl  = image.path;
+  const imageUrl = image.path;
 
   const product = new Product({
+    // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
     title: title,
     price: price,
     description: description,
@@ -72,8 +74,23 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      const error=new Error(err);
-      error.hhtpStatusCode=500;
+      // return res.status(500).render('admin/edit-product', {
+      //   pageTitle: 'Add Product',
+      //   path: '/admin/add-product',
+      //   editing: false,
+      //   hasError: true,
+      //   product: {
+      //     title: title,
+      //     imageUrl: imageUrl,
+      //     price: price,
+      //     description: description
+      //   },
+      //   errorMessage: 'Database operation failed, please try again.',
+      //   validationErrors: []
+      // });
+      // res.redirect('/500');
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -94,68 +111,70 @@ exports.getEditProduct = (req, res, next) => {
         path: '/admin/edit-product',
         editing: editMode,
         product: product,
-        hasError:false,
-        errorMessage:null,
-        validationErrors:[],
+        hasError: false,
+        errorMessage: null,
+        validationErrors: []
       });
     })
     .catch(err => {
-      const error=new Error(err);
-        error.hhtpStatusCode=500;
-        return next(error);
-      })
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
-  const errors=validationResult(req);
-  console.log(errors.array());
-  if(!errors.isEmpty()){
+
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Edit Product',
-      path: '/admin/add-product',
+      path: '/admin/edit-product',
       editing: true,
-      hasError:true,
+      hasError: true,
       product: {
-        title:updatedTitle,
-        imageUrl:updatedImageUrl,
-        price:updatedPrice,
-        description:updatedDesc,
-        _id:prodId
+        title: updatedTitle,
+        price: updatedPrice,
+        description: updatedDesc,
+        _id: prodId
       },
-      errorMessage:"Database operation failed,please try again.",
-      validationErrors:[],
-
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array()
     });
   }
 
   Product.findById(prodId)
     .then(product => {
-      if(product.userId.toString()!==req.user._id.toString()){
+      if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/');
       }
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
       });
     })
     .catch(err => {
-    const error=new Error(err);
-      error.hhtpStatusCode=500;
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
-    })
+    });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId:req.user._id})
+  Product.find({ userId: req.user._id })
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then(products => {
@@ -167,20 +186,27 @@ exports.getProducts = (req, res, next) => {
       });
     })
     .catch(err => {
-      const error=new Error(err);
-        error.hhtpStatusCode=500;
-        return next(error);
-      })
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.deleteProduct = (req, res, next) => {
   const prodId = req.params.productId;
-  Product.deleteOne({_id:prodId,userId:req.user._id})
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
-      res.status(200).json({message:'Success'});
+      res.status(200).json({ message: 'Success!' });
     })
     .catch(err => {
-      response.status(500).json({message:'Deleting product failed.'});
-    })
-    };
+      res.status(500).json({ message: 'Deleting product failed.' });
+    });
+};
